requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
anthropic==0.43.0
google-api-python-client==2.157.0
google-auth==2.37.0
google-auth-oauthlib==1.2.1
httpx==0.28.1
python-dotenv==1.0.1
pdfplumber==0.11.4
